add unit test for core item-base recommendation algorithm
