+++ .gitignore
@@ -6,4 +6,3 @@ data/
 default.conf
 nginx/
 nohup.out
-unit_test/
